pyproject.toml
[project]
name = "nepali-news-rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "faiss-cpu>=1.13.2",
    "groq>=1.1.2",
    "huggingface-hub<1.0.0",
    "langchain>=1.2.7",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "langdetect>=1.0.9",
    "llama-cpp-python>=0.3.0",
    "ollama>=0.6.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.53.1",
    "tqdm>=4.67.2",
    "transformers<5.0.0",
]
